Move appelflap route registration (mostly) out of sw.js
